backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
alpaca-py==0.8.2
pandas<2.0.0,>=1.3.5
python-dotenv==1.0.0
pytz==2023.3
pydantic<2.0.0,>=1.9.0
